Added specs for private method #neighborhood
This is a bit controversial – should you test private methods or not? I group them together in a describe block, so they're easy to distinguish.

The reason I include them is that a lot of the logic is passed down and broken up into private methods. To easy the implementation of them with TDD, I write tests for them.

After analyzing the problem, I find that I want a list of survivors –> this becomes the #survivors method. Survivors are those with a certain number of neighbors -> I need a #neighbors method. Neighbors are all the cells in the neighborhood, except the cell itself –> #neighborhood is needed, and can be calculated with some simple arithmetic.